+++ src/java/com/github/lucasgueiros/ifuwhist/partida/bolsa/Bolsa.java
@@ -6,8 +6,8 @@
 package com.github.lucasgueiros.ifuwhist.partida.bolsa;
 
 import com.github.lucasgueiros.ifuwhist.mesa.Posicao;
-import com.github.lucasgueiros.ifuwhist.partida.cartas.Carta;
-import com.github.lucasgueiros.ifuwhist.partida.cartas.Naipe;
+import com.github.lucasgueiros.ifuwhist.partida.vaza.Carta;
+import com.github.lucasgueiros.ifuwhist.partida.vaza.Naipe;
 import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
