+++ src/java/com/github/lucasgueiros/ifuwhist/partida/bolsa/EmbaralhadorSimples.java
@@ -6,10 +6,10 @@
 package com.github.lucasgueiros.ifuwhist.partida.bolsa;
 
 import com.github.lucasgueiros.ifuwhist.mesa.Posicao;
-import com.github.lucasgueiros.ifuwhist.partida.cartas.Carta;
-import com.github.lucasgueiros.ifuwhist.partida.cartas.ComparaCartas;
-import com.github.lucasgueiros.ifuwhist.partida.cartas.Naipe;
-import com.github.lucasgueiros.ifuwhist.partida.cartas.Simbolo;
+import com.github.lucasgueiros.ifuwhist.partida.vaza.Carta;
+import com.github.lucasgueiros.ifuwhist.partida.vaza.ComparaCartas;
+import com.github.lucasgueiros.ifuwhist.partida.vaza.Naipe;
+import com.github.lucasgueiros.ifuwhist.partida.vaza.Simbolo;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedList;
