+++ src/java/com/github/lucasgueiros/ifuwhist/partida/eventos/ListenerPartida.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.github.lucasgueiros.ifuwhist.partida;
+package com.github.lucasgueiros.ifuwhist.partida.eventos;
 
 /**
  *
