+++ src/java/com/github/lucasgueiros/ifuwhist/partida/vaza/Naipe.java
@@ -4,7 +4,7 @@
  * and open the template in the editor.
  */
 
-package com.github.lucasgueiros.ifuwhist.partida.cartas;
+package com.github.lucasgueiros.ifuwhist.partida.vaza;
 
 /**
  *
